requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt
dlib
imageio==2.6
numpy
opencv-python
tensorboard
tensorflow-gpu
imutils
streamlit
scipy
